Move control service and remote work to separate files
